@@ -18,14 +18,23 @@ pub fn elem(stream: TokenStream, body: syn::ItemStruct) -> Result<TokenStream> {
 
 /// Details about an element.
 struct Elem {
+    /// The element's name as exposed to Typst.
     name: String,
+    /// The element's title case name.
     title: String,
+    /// Whether this element has an associated scope defined by the `#[scope]` macro.
     scope: bool,
+    /// A list of alternate search terms for this element.
     keywords: Vec<String>,
+    /// The documentation for this element as a string.
     docs: String,
+    /// The element's visibility.
     vis: syn::Visibility,
+    /// The struct name for this element given in Rust.
     ident: Ident,
+    /// The list of capabilities for this element.
     capabilities: Vec<Ident>,
+    /// The fields of this element.
     fields: Vec<Field>,
 }
 
@@ -97,30 +106,59 @@ impl Elem {
     }
 }
 
+/// A field of an [element definition][`Elem`].
 struct Field {
+    /// The name of this field.
     ident: Ident,
+    /// The identifier `{ident}_in`.
     ident_in: Ident,
+    /// The identifier `with_{ident}`.
     with_ident: Ident,
+    /// The identifier `push_{ident}`.
     push_ident: Ident,
+    /// The identifier `set_{ident}`.
     set_ident: Ident,
+    /// The upper camel-case version of `ident`, used for the enum variant name.
     enum_ident: Ident,
+    /// The all-caps snake-case version of `ident`, used for the constant name.
     const_ident: Ident,
+    /// The visibility of this field.
     vis: syn::Visibility,
+    /// The type of this field.
     ty: syn::Type,
+    /// The type returned by accessor methods for this field.
+    ///
+    /// Usually, this is the same as `ty`, but this might be different
+    /// if this field has a `#[resolve]` attribute.
     output: syn::Type,
+    /// The field's identifier as exposed to Typst.
     name: String,
+    /// The documentation for this field as a string.
     docs: String,
+    /// Whether this field is positional (as opposed to named).
     positional: bool,
+    /// Whether this field is required.
     required: bool,
+    /// Whether this field is variadic; that is, has its values
+    /// taken from a variable number of arguments.
     variadic: bool,
+    /// Whether this field has a `#[resolve]` attribute.
     resolve: bool,
+    /// Whether this field has a `#[fold]` attribute.
     fold: bool,
+    /// Whether this field is excluded from documentation.
     internal: bool,
+    /// Whether this field exists only in documentation.
     external: bool,
+    /// Whether this field has a `#[borrowed]` attribute.
     borrowed: bool,
+    /// Whether this field has a `#[ghost]` attribute.
     ghost: bool,
+    /// Whether this field has a `#[synthesized]` attribute.
     synthesized: bool,
+    /// The contents of the `#[parse({..})]` attribute, if any.
     parse: Option<BlockWithReturn>,
+    /// The contents of the `#[default(..)]` attribute, if any.
     default: Option<syn::Expr>,
 }
 
@@ -276,6 +314,7 @@ fn create(element: &Elem) -> Result<TokenStream> {
     let fields_impl = create_fields_impl(element);
     let repr_impl = element.cannot("Repr").then(|| create_repr_impl(element));
     let locatable_impl = element.can("Locatable").then(|| create_locatable_impl(element));
+    let mathy_impl = element.can("Mathy").then(|| create_mathy_impl(element));
     let into_value_impl = create_into_value_impl(element);
 
     // We use a const block to create an anonymous scope, as to not leak any
@@ -295,6 +334,7 @@ fn create(element: &Elem) -> Result<TokenStream> {
             #partial_eq_impl
             #repr_impl
             #locatable_impl
+            #mathy_impl
             #into_value_impl
         };
     })
@@ -352,13 +392,13 @@ fn create_fields_enum(element: &Elem) -> TokenStream {
         }
 
         impl ::std::convert::TryFrom<u8> for Fields {
-            type Error = ();
+            type Error = #foundations::FieldAccessError;
 
             fn try_from(value: u8) -> Result<Self, Self::Error> {
                 #(const #consts: u8 = Fields::#variants as u8;)*
                 match value {
                     #(#consts => Ok(Self::#variants),)*
-                    _ => Err(()),
+                    _ => Err(#foundations::FieldAccessError::Internal),
                 }
             }
         }
@@ -392,8 +432,8 @@ fn create_default_static(field: &Field) -> TokenStream {
     };
 
     quote! {
-        static #const_ident: ::once_cell::sync::Lazy<#ty> =
-            ::once_cell::sync::Lazy::new(#init);
+        static #const_ident: ::std::sync::LazyLock<#ty> =
+            ::std::sync::LazyLock::new(#init);
     }
 }
 
@@ -594,7 +634,7 @@ fn create_native_elem_impl(element: &Elem) -> TokenStream {
     let Elem { name, ident, title, scope, keywords, docs, .. } = element;
 
     let local_name = if element.can("LocalName") {
-        quote! { Some(<#foundations::Packed<#ident> as ::typst::text::LocalName>::local_name) }
+        quote! { Some(<#foundations::Packed<#ident> as ::typst_library::text::LocalName>::local_name) }
     } else {
         quote! { None }
     };
@@ -620,8 +660,8 @@ fn create_native_elem_impl(element: &Elem) -> TokenStream {
             field_name: |id| id.try_into().ok().map(Fields::to_str),
             field_from_styles: <#ident as #foundations::Fields>::field_from_styles,
             local_name: #local_name,
-            scope: #foundations::Lazy::new(|| #scope),
-            params: #foundations::Lazy::new(|| ::std::vec![#(#params),*])
+            scope: ::std::sync::LazyLock::new(|| #scope),
+            params: ::std::sync::LazyLock::new(|| ::std::vec![#(#params),*])
         }
     };
 
@@ -723,9 +763,9 @@ fn create_construct_impl(element: &Elem) -> TokenStream {
     quote! {
         impl #foundations::Construct for #ident {
             fn construct(
-                engine: &mut ::typst::engine::Engine,
+                engine: &mut ::typst_library::engine::Engine,
                 args: &mut #foundations::Args,
-            ) -> ::typst::diag::SourceResult<#foundations::Content> {
+            ) -> ::typst_library::diag::SourceResult<#foundations::Content> {
                 #(#setup)*
                 Ok(#foundations::Content::new(Self { #(#fields),* }))
             }
@@ -750,9 +790,9 @@ fn create_set_impl(element: &Elem) -> TokenStream {
     quote! {
         impl #foundations::Set for #ident {
             fn set(
-                engine: &mut ::typst::engine::Engine,
+                engine: &mut ::typst_library::engine::Engine,
                 args: &mut #foundations::Args,
-            ) -> ::typst::diag::SourceResult<#foundations::Styles> {
+            ) -> ::typst_library::diag::SourceResult<#foundations::Styles> {
                 let mut styles = #foundations::Styles::new();
                 #(#handlers)*
                 Ok(styles)
@@ -799,7 +839,7 @@ fn create_capable_impl(element: &Elem) -> TokenStream {
                 // Safety: The vtable function doesn't require initialized
                 // data, so it's fine to use a dangling pointer.
                 return Some(unsafe {
-                    ::typst::util::fat::vtable(dangling as *const dyn #capability)
+                    ::typst_utils::fat::vtable(dangling as *const dyn #capability)
                 });
             }
         }
@@ -807,7 +847,7 @@ fn create_capable_impl(element: &Elem) -> TokenStream {
 
     quote! {
         unsafe impl #foundations::Capable for #ident {
-            fn vtable(capability: ::std::any::TypeId) -> ::std::option::Option<*const ()> {
+            fn vtable(capability: ::std::any::TypeId) -> ::std::option::Option<::std::ptr::NonNull<()>> {
                 let dangling = ::std::ptr::NonNull::<#foundations::Packed<#ident>>::dangling().as_ptr();
                 #(#checks)*
                 None
@@ -839,9 +879,9 @@ fn create_fields_impl(element: &Elem) -> TokenStream {
         let Field { enum_ident, ident, .. } = field;
 
         let expr = if field.required {
-            quote! { Some(#into_value(self.#ident.clone())) }
+            quote! { Ok(#into_value(self.#ident.clone())) }
         } else {
-            quote! { self.#ident.clone().map(#into_value) }
+            quote! { self.#ident.clone().map(#into_value).ok_or(#foundations::FieldAccessError::Unset) }
         };
 
         quote! { Fields::#enum_ident => #expr }
@@ -852,9 +892,9 @@ fn create_fields_impl(element: &Elem) -> TokenStream {
         let Field { enum_ident, ident, .. } = field;
 
         let expr = if field.required {
-            quote! { Some(#into_value(self.#ident.clone())) }
+            quote! { Ok(#into_value(self.#ident.clone())) }
         } else if field.synthesized {
-            quote! { self.#ident.clone().map(#into_value) }
+            quote! { self.#ident.clone().map(#into_value).ok_or(#foundations::FieldAccessError::Unset) }
         } else {
             let value = create_style_chain_access(
                 field,
@@ -862,7 +902,7 @@ fn create_fields_impl(element: &Elem) -> TokenStream {
                 if field.ghost { quote!(None) } else { quote!(self.#ident.as_ref()) },
             );
 
-            quote! { Some(#into_value(#value)) }
+            quote! { Ok(#into_value(#value)) }
         };
 
         quote! { Fields::#enum_ident => #expr }
@@ -873,10 +913,10 @@ fn create_fields_impl(element: &Elem) -> TokenStream {
         let Field { enum_ident, .. } = field;
 
         let expr = if field.required || field.synthesized {
-            quote! { None }
+            quote! { Err(#foundations::FieldAccessError::Unknown) }
         } else {
             let value = create_style_chain_access(field, false, quote!(None));
-            quote! { Some(#into_value(#value)) }
+            quote! { Ok(#into_value(#value)) }
         };
 
         quote! { Fields::#enum_ident => #expr }
@@ -924,6 +964,10 @@ fn create_fields_impl(element: &Elem) -> TokenStream {
 
     let Elem { ident, .. } = element;
 
+    let result = quote! {
+        Result<#foundations::Value, #foundations::FieldAccessError>
+    };
+
     quote! {
         impl #foundations::Fields for #ident {
             type Enum = Fields;
@@ -939,27 +983,33 @@ fn create_fields_impl(element: &Elem) -> TokenStream {
                 }
             }
 
-            fn field(&self, id: u8) -> Option<#foundations::Value> {
-                let id = Fields::try_from(id).ok()?;
+            fn field(&self, id: u8) -> #result {
+                let id = Fields::try_from(id)?;
                 match id {
                     #(#field_arms,)*
-                    _ => None,
+                    // This arm might be reached if someone tries to access an
+                    // internal field.
+                    _ => Err(#foundations::FieldAccessError::Unknown),
                 }
             }
 
-            fn field_with_styles(&self, id: u8, styles: #foundations::StyleChain) -> Option<#foundations::Value> {
-                let id = Fields::try_from(id).ok()?;
+            fn field_with_styles(&self, id: u8, styles: #foundations::StyleChain) -> #result {
+                let id = Fields::try_from(id)?;
                 match id {
                     #(#field_with_styles_arms,)*
-                    _ => None,
+                    // This arm might be reached if someone tries to access an
+                    // internal field.
+                    _ => Err(#foundations::FieldAccessError::Unknown),
                 }
             }
 
-            fn field_from_styles(id: u8, styles: #foundations::StyleChain) -> Option<#foundations::Value> {
-                let id = Fields::try_from(id).ok()?;
+            fn field_from_styles(id: u8, styles: #foundations::StyleChain) -> #result {
+                let id = Fields::try_from(id)?;
                 match id {
                     #(#field_from_styles_arms,)*
-                    _ => None,
+                    // This arm might be reached if someone tries to access an
+                    // internal field.
+                    _ => Err(#foundations::FieldAccessError::Unknown),
                 }
             }
 
@@ -999,7 +1049,13 @@ fn create_repr_impl(element: &Elem) -> TokenStream {
 /// Creates the element's `Locatable` implementation.
 fn create_locatable_impl(element: &Elem) -> TokenStream {
     let ident = &element.ident;
-    quote! { impl ::typst::introspection::Locatable for #foundations::Packed<#ident> {} }
+    quote! { impl ::typst_library::introspection::Locatable for #foundations::Packed<#ident> {} }
+}
+
+/// Creates the element's `Mathy` implementation.
+fn create_mathy_impl(element: &Elem) -> TokenStream {
+    let ident = &element.ident;
+    quote! { impl ::typst_library::math::Mathy for #foundations::Packed<#ident> {} }
 }
 
 /// Creates the element's `IntoValue` implementation.
