@@ -1,18 +1,48 @@
 use std::fmt::{self, Display, Formatter};
+use std::num::NonZeroUsize;
+use std::ops::RangeInclusive;
 use std::path::PathBuf;
+use std::str::FromStr;
 
 use chrono::{DateTime, Utc};
-use clap::builder::ValueParser;
-use clap::{ArgAction, Args, ColorChoice, Parser, Subcommand, ValueEnum};
+use clap::builder::{TypedValueParser, ValueParser};
+use clap::{ArgAction, Args, ColorChoice, Parser, Subcommand, ValueEnum, ValueHint};
 use semver::Version;
 
 /// The character typically used to separate path components
 /// in environment variables.
 const ENV_PATH_SEP: char = if cfg!(windows) { ';' } else { ':' };
 
-/// The Typst compiler.
+/// The overall structure of the help.
+#[rustfmt::skip]
+const HELP_TEMPLATE: &str = "\
+Typst {version}
+
+{usage-heading} {usage}
+
+{all-args}{after-help}\
+";
+
+/// Adds a list of useful links after the normal help.
+#[rustfmt::skip]
+const AFTER_HELP: &str = color_print::cstr!("\
+<s><u>Resources:</></>
+  <s>Tutorial:</>                 https://typst.app/docs/tutorial/
+  <s>Reference documentation:</>  https://typst.app/docs/reference/
+  <s>Templates & Packages:</>     https://typst.app/universe/
+  <s>Forum for questions:</>      https://forum.typst.app/
+");
+
+/// The Typst compiler
 #[derive(Debug, Clone, Parser)]
-#[clap(name = "typst", version = crate::typst_version(), author)]
+#[clap(
+    name = "typst",
+    version = crate::typst_version(),
+    author,
+    help_template = HELP_TEMPLATE,
+    after_help = AFTER_HELP,
+    max_term_width = 80,
+)]
 pub struct CliArguments {
     /// The command to run
     #[command(subcommand)]
@@ -57,7 +87,7 @@ pub enum Command {
     Fonts(FontsCommand),
 
     /// Self update the Typst CLI
-    #[cfg_attr(not(feature = "self-update"), doc = " (disabled)")]
+    #[cfg_attr(not(feature = "self-update"), clap(hide = true))]
     Update(UpdateCommand),
 }
 
@@ -68,9 +98,30 @@ pub struct CompileCommand {
     #[clap(flatten)]
     pub common: SharedArgs,
 
-    /// Path to output file (PDF, PNG, or SVG), use `-` to write output to stdout
-    #[clap(required_if_eq("input", "-"), value_parser = ValueParser::new(output_value_parser))]
+    /// Path to output file (PDF, PNG or SVG). Use `-` to write output to stdout.
+    ///
+    /// For output formats emitting one file per page (PNG & SVG), a page number template
+    /// must be present if the source document renders to multiple pages. Use `{p}` for page
+    /// numbers, `{0p}` for zero padded page numbers and `{t}` for page count. For example,
+    /// `page-{0p}-of-{t}.png` creates `page-01-of-10.png`, `page-02-of-10.png` and so on.
+    #[clap(
+        required_if_eq("input", "-"),
+        value_parser = make_output_value_parser(),
+        value_hint = ValueHint::FilePath,
+    )]
     pub output: Option<Output>,
+
+    /// Which pages to export. When unspecified, all document pages are exported.
+    ///
+    /// Pages to export are separated by commas, and can be either simple page
+    /// numbers (e.g. '2,5' to export only pages 2 and 5) or page ranges
+    /// (e.g. '2,3-6,8-' to export page 2, pages 3 to 6 (inclusive), page 8 and
+    /// any pages after it).
+    ///
+    /// Page numbers are one-indexed and correspond to real page numbers in the
+    /// document (therefore not being affected by the document's page counter).
+    #[arg(long = "pages", value_delimiter = ',')]
+    pub pages: Option<Vec<PageRangeArgument>>,
 
     /// Output a Makefile rule describing the current compilation
     #[clap(long = "make-deps", value_name = "PATH")]
@@ -80,9 +131,11 @@ pub struct CompileCommand {
     #[arg(long = "format", short = 'f')]
     pub format: Option<OutputFormat>,
 
-    /// Opens the output file using the default viewer after compilation.
-    /// Ignored if output is stdout
-    #[arg(long = "open")]
+    /// Opens the output file with the default viewer or a specific program after
+    /// compilation
+    ///
+    /// Ignored if output is stdout.
+    #[arg(long = "open", value_name = "VIEWER")]
     pub open: Option<Option<String>>,
 
     /// The PPI (pixels per inch) to use for PNG export
@@ -96,6 +149,23 @@ pub struct CompileCommand {
     /// apart from file names and line numbers.
     #[arg(long = "timings", value_name = "OUTPUT_JSON")]
     pub timings: Option<Option<PathBuf>>,
+
+    /// One (or multiple comma-separated) PDF standards that Typst will enforce
+    /// conformance with.
+    #[arg(long = "pdf-standard", value_delimiter = ',')]
+    pub pdf_standard: Vec<PdfStandard>,
+}
+
+/// A PDF standard that Typst can enforce conformance with.
+#[derive(Debug, Copy, Clone, Eq, PartialEq, ValueEnum)]
+#[allow(non_camel_case_types)]
+pub enum PdfStandard {
+    /// PDF 1.7.
+    #[value(name = "1.7")]
+    V_1_7,
+    /// PDF/A-2b.
+    #[value(name = "a-2b")]
+    A_2b,
 }
 
 /// Initializes a new project from a template
@@ -111,6 +181,10 @@ pub struct InitCommand {
 
     /// The project directory, defaults to the template's name
     pub dir: Option<String>,
+
+    /// Arguments related to storage of packages in the system
+    #[clap(flatten)]
+    pub package_storage_args: PackageStorageArgs,
 }
 
 /// Processes an input file to extract provided metadata
@@ -134,6 +208,12 @@ pub struct QueryCommand {
     /// The format to serialize in
     #[clap(long = "format", default_value = "json")]
     pub format: SerializationFormat,
+
+    /// Whether to pretty-print the serialized output.
+    ///
+    /// Only applies to JSON format.
+    #[clap(long)]
+    pub pretty: bool,
 }
 
 // Output file format for query command
@@ -146,8 +226,8 @@ pub enum SerializationFormat {
 /// Common arguments of compile, watch, and query.
 #[derive(Debug, Clone, Args)]
 pub struct SharedArgs {
-    /// Path to input Typst file, use `-` to read input from stdin
-    #[clap(value_parser = input_value_parser)]
+    /// Path to input Typst file. Use `-` to read input from stdin
+    #[clap(value_parser = make_input_value_parser(), value_hint = ValueHint::FilePath)]
     pub input: Input,
 
     /// Configures the project root (for absolute paths)
@@ -163,14 +243,9 @@ pub struct SharedArgs {
     )]
     pub inputs: Vec<(String, String)>,
 
-    /// Adds additional directories to search for fonts
-    #[clap(
-        long = "font-path",
-        env = "TYPST_FONT_PATHS",
-        value_name = "DIR",
-        value_delimiter = ENV_PATH_SEP,
-    )]
-    pub font_paths: Vec<PathBuf>,
+    /// Common font arguments
+    #[clap(flatten)]
+    pub font_args: FontArgs,
 
     /// The document's creation date formatted as a UNIX timestamp.
     ///
@@ -190,6 +265,31 @@ pub struct SharedArgs {
         value_parser = clap::value_parser!(DiagnosticFormat)
     )]
     pub diagnostic_format: DiagnosticFormat,
+
+    /// Arguments related to storage of packages in the system
+    #[clap(flatten)]
+    pub package_storage_args: PackageStorageArgs,
+
+    /// Number of parallel jobs spawned during compilation,
+    /// defaults to number of CPUs. Setting it to 1 disables parallelism.
+    #[clap(long, short)]
+    pub jobs: Option<usize>,
+}
+
+/// Arguments related to where packages are stored in the system.
+#[derive(Debug, Clone, Args)]
+pub struct PackageStorageArgs {
+    /// Custom path to local packages, defaults to system-dependent location
+    #[clap(long = "package-path", env = "TYPST_PACKAGE_PATH", value_name = "DIR")]
+    pub package_path: Option<PathBuf>,
+
+    /// Custom path to package cache, defaults to system-dependent location
+    #[clap(
+        long = "package-cache-path",
+        env = "TYPST_PACKAGE_CACHE_PATH",
+        value_name = "DIR"
+    )]
+    pub package_cache_path: Option<PathBuf>,
 }
 
 /// Parses a UNIX timestamp according to <https://reproducible-builds.org/specs/source-date-epoch/>
@@ -229,26 +329,30 @@ impl Display for Output {
 }
 
 /// The clap value parser used by `SharedArgs.input`
-fn input_value_parser(value: &str) -> Result<Input, clap::error::Error> {
-    if value.is_empty() {
-        Err(clap::Error::new(clap::error::ErrorKind::InvalidValue))
-    } else if value == "-" {
-        Ok(Input::Stdin)
-    } else {
-        Ok(Input::Path(value.into()))
-    }
+fn make_input_value_parser() -> impl TypedValueParser<Value = Input> {
+    clap::builder::OsStringValueParser::new().try_map(|value| {
+        if value.is_empty() {
+            Err(clap::Error::new(clap::error::ErrorKind::InvalidValue))
+        } else if value == "-" {
+            Ok(Input::Stdin)
+        } else {
+            Ok(Input::Path(value.into()))
+        }
+    })
 }
 
 /// The clap value parser used by `CompileCommand.output`
-fn output_value_parser(value: &str) -> Result<Output, clap::error::Error> {
-    // Empty value also handled by clap for `Option<Output>`
-    if value.is_empty() {
-        Err(clap::Error::new(clap::error::ErrorKind::InvalidValue))
-    } else if value == "-" {
-        Ok(Output::Stdout)
-    } else {
-        Ok(Output::Path(value.into()))
-    }
+fn make_output_value_parser() -> impl TypedValueParser<Value = Output> {
+    clap::builder::OsStringValueParser::new().try_map(|value| {
+        // Empty value also handled by clap for `Option<Output>`
+        if value.is_empty() {
+            Err(clap::Error::new(clap::error::ErrorKind::InvalidValue))
+        } else if value == "-" {
+            Ok(Output::Stdout)
+        } else {
+            Ok(Output::Path(value.into()))
+        }
+    })
 }
 
 /// Parses key/value pairs split by the first equal sign.
@@ -267,10 +371,74 @@ fn parse_input_pair(raw: &str) -> Result<(String, String), String> {
     Ok((key, val))
 }
 
+/// Implements parsing of page ranges (`1-3`, `4`, `5-`, `-2`), used by the
+/// `CompileCommand.pages` argument, through the `FromStr` trait instead of
+/// a value parser, in order to generate better errors.
+///
+/// See also: https://github.com/clap-rs/clap/issues/5065
+#[derive(Debug, Clone)]
+pub struct PageRangeArgument(RangeInclusive<Option<NonZeroUsize>>);
+
+impl PageRangeArgument {
+    pub fn to_range(&self) -> RangeInclusive<Option<NonZeroUsize>> {
+        self.0.clone()
+    }
+}
+
+impl FromStr for PageRangeArgument {
+    type Err = &'static str;
+
+    fn from_str(value: &str) -> Result<Self, Self::Err> {
+        match value.split('-').map(str::trim).collect::<Vec<_>>().as_slice() {
+            [] | [""] => Err("page export range must not be empty"),
+            [single_page] => {
+                let page_number = parse_page_number(single_page)?;
+                Ok(PageRangeArgument(Some(page_number)..=Some(page_number)))
+            }
+            ["", ""] => Err("page export range must have start or end"),
+            [start, ""] => Ok(PageRangeArgument(Some(parse_page_number(start)?)..=None)),
+            ["", end] => Ok(PageRangeArgument(None..=Some(parse_page_number(end)?))),
+            [start, end] => {
+                let start = parse_page_number(start)?;
+                let end = parse_page_number(end)?;
+                if start > end {
+                    Err("page export range must end at a page after the start")
+                } else {
+                    Ok(PageRangeArgument(Some(start)..=Some(end)))
+                }
+            }
+            [_, _, _, ..] => Err("page export range must have a single hyphen"),
+        }
+    }
+}
+
+fn parse_page_number(value: &str) -> Result<NonZeroUsize, &'static str> {
+    if value == "0" {
+        Err("page numbers start at one")
+    } else {
+        NonZeroUsize::from_str(value).map_err(|_| "not a valid page number")
+    }
+}
+
 /// Lists all discovered fonts in system and custom font paths
 #[derive(Debug, Clone, Parser)]
 pub struct FontsCommand {
-    /// Adds additional directories to search for fonts
+    /// Common font arguments
+    #[clap(flatten)]
+    pub font_args: FontArgs,
+
+    /// Also lists style variants of each font family
+    #[arg(long)]
+    pub variants: bool,
+}
+
+/// Common arguments to customize available fonts
+#[derive(Debug, Clone, Parser)]
+pub struct FontArgs {
+    /// Adds additional directories that are recursively searched for fonts
+    ///
+    /// If multiple paths are specified, they are separated by the system's path
+    /// separator (`:` on Unix-like systems and `;` on Windows).
     #[clap(
         long = "font-path",
         env = "TYPST_FONT_PATHS",
@@ -279,9 +447,10 @@ pub struct FontsCommand {
     )]
     pub font_paths: Vec<PathBuf>,
 
-    /// Also lists style variants of each font family
+    /// Ensures system fonts won't be searched, unless explicitly included via
+    /// `--font-path`
     #[arg(long)]
-    pub variants: bool,
+    pub ignore_system_fonts: bool,
 }
 
 /// Which format to use for diagnostics.
@@ -312,8 +481,18 @@ pub struct UpdateCommand {
 
     /// Reverts to the version from before the last update (only possible if
     /// `typst update` has previously ran)
-    #[clap(long, default_value_t = false, exclusive = true)]
+    #[clap(
+        long,
+        default_value_t = false,
+        conflicts_with = "version",
+        conflicts_with = "force"
+    )]
     pub revert: bool,
+
+    /// Custom path to the backup file created on update and used by `--revert`,
+    /// defaults to system-dependent location
+    #[clap(long = "backup-path", env = "TYPST_UPDATE_BACKUP_PATH", value_name = "FILE")]
+    pub backup_path: Option<PathBuf>,
 }
 
 /// Which format to use for the generated output file.
