@@ -2,6 +2,7 @@ mod args;
 mod compile;
 mod download;
 mod fonts;
+mod greet;
 mod init;
 mod package;
 mod query;
@@ -15,11 +16,13 @@ mod world;
 use std::cell::Cell;
 use std::io::{self, Write};
 use std::process::ExitCode;
+use std::sync::LazyLock;
 
+use clap::error::ErrorKind;
 use clap::Parser;
 use codespan_reporting::term;
 use codespan_reporting::term::termcolor::WriteColor;
-use once_cell::sync::Lazy;
+use typst::diag::HintedStrResult;
 
 use crate::args::{CliArguments, Command};
 use crate::timings::Timer;
@@ -29,28 +32,42 @@ thread_local! {
     static EXIT: Cell<ExitCode> = const { Cell::new(ExitCode::SUCCESS) };
 }
 
-/// The parsed commandline arguments.
-static ARGS: Lazy<CliArguments> = Lazy::new(CliArguments::parse);
+/// The parsed command line arguments.
+static ARGS: LazyLock<CliArguments> = LazyLock::new(|| {
+    CliArguments::try_parse().unwrap_or_else(|error| {
+        if error.kind() == ErrorKind::DisplayHelpOnMissingArgumentOrSubcommand {
+            crate::greet::greet();
+        }
+        error.exit();
+    })
+});
 
 /// Entry point.
 fn main() -> ExitCode {
-    let timer = Timer::new(&ARGS);
-
-    let res = match &ARGS.command {
-        Command::Compile(command) => crate::compile::compile(timer, command.clone()),
-        Command::Watch(command) => crate::watch::watch(timer, command.clone()),
-        Command::Init(command) => crate::init::init(command),
-        Command::Query(command) => crate::query::query(command),
-        Command::Fonts(command) => crate::fonts::fonts(command),
-        Command::Update(command) => crate::update::update(command),
-    };
+    let res = dispatch();
 
     if let Err(msg) = res {
         set_failed();
-        print_error(&msg).expect("failed to print error");
+        print_error(msg.message()).expect("failed to print error");
     }
 
     EXIT.with(|cell| cell.get())
+}
+
+/// Execute the requested command.
+fn dispatch() -> HintedStrResult<()> {
+    let timer = Timer::new(&ARGS);
+
+    match &ARGS.command {
+        Command::Compile(command) => crate::compile::compile(timer, command.clone())?,
+        Command::Watch(command) => crate::watch::watch(timer, command.clone())?,
+        Command::Init(command) => crate::init::init(command)?,
+        Command::Query(command) => crate::query::query(command)?,
+        Command::Fonts(command) => crate::fonts::fonts(command),
+        Command::Update(command) => crate::update::update(command)?,
+    }
+
+    Ok(())
 }
 
 /// Ensure a failure exit code.
@@ -77,8 +94,9 @@ fn print_error(msg: &str) -> io::Result<()> {
 
 #[cfg(not(feature = "self-update"))]
 mod update {
-    use crate::args::UpdateCommand;
     use typst::diag::{bail, StrResult};
+
+    use crate::args::UpdateCommand;
 
     pub fn update(_: &UpdateCommand) -> StrResult<()> {
         bail!(
