@@ -1,12 +1,12 @@
 use comemo::Track;
 use ecow::{eco_format, EcoString};
 use serde::Serialize;
-use typst::diag::{bail, StrResult};
-use typst::eval::{eval_string, EvalMode, Tracer};
+use typst::diag::{bail, HintedStrResult, StrResult, Warned};
 use typst::foundations::{Content, IntoValue, LocatableSelector, Scope};
 use typst::model::Document;
 use typst::syntax::Span;
 use typst::World;
+use typst_eval::{eval_string, EvalMode};
 
 use crate::args::{QueryCommand, SerializationFormat};
 use crate::compile::print_diagnostics;
@@ -14,18 +14,16 @@ use crate::set_failed;
 use crate::world::SystemWorld;
 
 /// Execute a query command.
-pub fn query(command: &QueryCommand) -> StrResult<()> {
+pub fn query(command: &QueryCommand) -> HintedStrResult<()> {
     let mut world = SystemWorld::new(&command.common)?;
 
     // Reset everything and ensure that the main file is present.
     world.reset();
     world.source(world.main()).map_err(|err| err.to_string())?;
 
-    let mut tracer = Tracer::new();
-    let result = typst::compile(&world, &mut tracer);
-    let warnings = tracer.warnings();
+    let Warned { output, warnings } = typst::compile(&world);
 
-    match result {
+    match output {
         // Retrieve and print query results.
         Ok(document) => {
             let data = retrieve(&world, command, &document)?;
@@ -56,8 +54,9 @@ fn retrieve(
     world: &dyn World,
     command: &QueryCommand,
     document: &Document,
-) -> StrResult<Vec<Content>> {
+) -> HintedStrResult<Vec<Content>> {
     let selector = eval_string(
+        &typst::ROUTINES,
         world.track(),
         &command.selector,
         Span::detached(),
@@ -90,7 +89,7 @@ fn format(elements: Vec<Content>, command: &QueryCommand) -> StrResult<String> {
     let mapped: Vec<_> = elements
         .into_iter()
         .filter_map(|c| match &command.field {
-            Some(field) => c.get_by_name(field),
+            Some(field) => c.get_by_name(field).ok(),
             _ => Some(c.into_value()),
         })
         .collect();
@@ -99,20 +98,28 @@ fn format(elements: Vec<Content>, command: &QueryCommand) -> StrResult<String> {
         let Some(value) = mapped.first() else {
             bail!("no such field found for element");
         };
-        serialize(value, command.format)
+        serialize(value, command.format, command.pretty)
     } else {
-        serialize(&mapped, command.format)
+        serialize(&mapped, command.format, command.pretty)
     }
 }
 
 /// Serialize data to the output format.
-fn serialize(data: &impl Serialize, format: SerializationFormat) -> StrResult<String> {
+fn serialize(
+    data: &impl Serialize,
+    format: SerializationFormat,
+    pretty: bool,
+) -> StrResult<String> {
     match format {
         SerializationFormat::Json => {
-            serde_json::to_string_pretty(data).map_err(|e| eco_format!("{e}"))
+            if pretty {
+                serde_json::to_string_pretty(data).map_err(|e| eco_format!("{e}"))
+            } else {
+                serde_json::to_string(data).map_err(|e| eco_format!("{e}"))
+            }
         }
         SerializationFormat::Yaml => {
-            serde_yaml::to_string(&data).map_err(|e| eco_format!("{e}"))
+            serde_yaml::to_string(data).map_err(|e| eco_format!("{e}"))
         }
     }
 }
