@@ -1,15 +1,16 @@
 use std::io::{self, IsTerminal, Write};
 
 use codespan_reporting::term::termcolor;
-use once_cell::sync::Lazy;
 use termcolor::{ColorChoice, WriteColor};
+use typst::utils::singleton;
 
 use crate::ARGS;
 
 /// Returns a handle to the optionally colored terminal output.
 pub fn out() -> TermOut {
-    static OUTPUT: Lazy<TermOutInner> = Lazy::new(TermOutInner::new);
-    TermOut { inner: &OUTPUT }
+    TermOut {
+        inner: singleton!(TermOutInner, TermOutInner::new()),
+    }
 }
 
 /// The stuff that has to be shared between instances of [`TermOut`].
@@ -58,7 +59,7 @@ impl TermOut {
         // We don't want to clear anything that is not a TTY.
         if self.inner.stream.supports_color() {
             // First, move the cursor up `lines` lines.
-            // Then, clear everything between between the cursor to end of screen.
+            // Then, clear everything between the cursor to end of screen.
             let mut stream = self.inner.stream.lock();
             write!(stream, "\x1B[1F\x1B[0J")?;
             stream.flush()?;
