@@ -108,7 +108,7 @@ pub fn cast(stream: TokenStream) -> Result<TokenStream> {
     let from_value = (!input.from_value.is_empty() || input.dynamic).then(|| {
         quote! {
             impl #foundations::FromValue for #ty {
-                fn from_value(value: #foundations::Value) -> ::typst::diag::StrResult<Self> {
+                fn from_value(value: #foundations::Value) -> ::typst_library::diag::HintedStrResult<Self> {
                     #from_value_body
                 }
             }
