@@ -1,23 +1,28 @@
-use std::collections::BTreeMap;
+use std::collections::{BTreeMap, HashMap};
+use std::hash::Hash;
 use std::sync::Arc;
 
 use ecow::{eco_format, EcoString};
 use pdf_writer::types::{CidFontType, FontFlags, SystemInfo, UnicodeCmap};
-use pdf_writer::writers::FontDescriptor;
-use pdf_writer::{Filter, Finish, Name, Rect, Str};
+use pdf_writer::writers::{FontDescriptor, WMode};
+use pdf_writer::{Chunk, Filter, Finish, Name, Rect, Ref, Str};
+use subsetter::GlyphRemapper;
 use ttf_parser::{name_id, GlyphId, Tag};
-use typst::layout::{Abs, Em, Ratio, Transform};
-use typst::text::Font;
-use typst::util::SliceExt;
-use unicode_properties::{GeneralCategory, UnicodeGeneralCategory};
+use typst_library::diag::{At, SourceResult};
+use typst_library::text::Font;
+use typst_syntax::Span;
+use typst_utils::SliceExt;
 
-use crate::page::{write_frame, PageContext};
-use crate::{deflate, AbsExt, EmExt, PdfContext};
+use crate::{deflate, EmExt, NameExt, PdfChunk, WithGlobalRefs};
 
 const CFF: Tag = Tag::from_bytes(b"CFF ");
 const CFF2: Tag = Tag::from_bytes(b"CFF2");
-const CMAP_NAME: Name = Name(b"Custom");
-const SYSTEM_INFO: SystemInfo = SystemInfo {
+
+const SUBSET_TAG_LEN: usize = 6;
+const IDENTITY_H: &str = "Identity-H";
+
+pub(crate) const CMAP_NAME: Name = Name(b"Custom");
+pub(crate) const SYSTEM_INFO: SystemInfo = SystemInfo {
     registry: Str(b"Adobe"),
     ordering: Str(b"Identity"),
     supplement: 0,
@@ -25,237 +30,143 @@ const SYSTEM_INFO: SystemInfo = SystemInfo {
 
 /// Embed all used fonts into the PDF.
 #[typst_macros::time(name = "write fonts")]
-pub(crate) fn write_fonts(ctx: &mut PdfContext) {
-    write_color_fonts(ctx);
-
-    for font in ctx.font_map.items() {
-        let type0_ref = ctx.alloc.bump();
-        let cid_ref = ctx.alloc.bump();
-        let descriptor_ref = ctx.alloc.bump();
-        let cmap_ref = ctx.alloc.bump();
-        let data_ref = ctx.alloc.bump();
-        ctx.font_refs.push(type0_ref);
-
-        let glyph_set = ctx.glyph_sets.get_mut(font).unwrap();
-        let ttf = font.ttf();
-
-        // Do we have a TrueType or CFF font?
-        //
-        // FIXME: CFF2 must be handled differently and requires PDF 2.0
-        // (or we have to convert it to CFF).
-        let is_cff = ttf
-            .raw_face()
-            .table(CFF)
-            .or_else(|| ttf.raw_face().table(CFF2))
-            .is_some();
-
-        let postscript_name = font
-            .find_name(name_id::POST_SCRIPT_NAME)
-            .unwrap_or_else(|| "unknown".to_string());
-
-        let subset_tag = subset_tag(glyph_set);
-        let base_font = eco_format!("{subset_tag}+{postscript_name}");
-        let base_font_type0 = if is_cff {
-            eco_format!("{base_font}-Identity-H")
-        } else {
-            base_font.clone()
-        };
-
-        // Write the base font object referencing the CID font.
-        ctx.pdf
-            .type0_font(type0_ref)
-            .base_font(Name(base_font_type0.as_bytes()))
-            .encoding_predefined(Name(b"Identity-H"))
-            .descendant_font(cid_ref)
-            .to_unicode(cmap_ref);
-
-        // Write the CID font referencing the font descriptor.
-        let mut cid = ctx.pdf.cid_font(cid_ref);
-        cid.subtype(if is_cff { CidFontType::Type0 } else { CidFontType::Type2 });
-        cid.base_font(Name(base_font.as_bytes()));
-        cid.system_info(SYSTEM_INFO);
-        cid.font_descriptor(descriptor_ref);
-        cid.default_width(0.0);
-        if !is_cff {
-            cid.cid_to_gid_map_predefined(Name(b"Identity"));
-        }
-
-        // Extract the widths of all glyphs.
-        let mut widths = vec![];
-        for gid in std::iter::once(0).chain(glyph_set.keys().copied()) {
-            let width = ttf.glyph_hor_advance(GlyphId(gid)).unwrap_or(0);
-            let units = font.to_em(width).to_font_units();
-            let cid = glyph_cid(font, gid);
-            if usize::from(cid) >= widths.len() {
-                widths.resize(usize::from(cid) + 1, 0.0);
-                widths[usize::from(cid)] = units;
-            }
-        }
-
-        // Write all non-zero glyph widths.
-        let mut first = 0;
-        let mut width_writer = cid.widths();
-        for (w, group) in widths.group_by_key(|&w| w) {
-            let end = first + group.len();
-            if w != 0.0 {
-                let last = end - 1;
-                width_writer.same(first as u16, last as u16, w);
-            }
-            first = end;
-        }
-
-        width_writer.finish();
-        cid.finish();
-
-        // Write the /ToUnicode character map, which maps glyph ids back to
-        // unicode codepoints to enable copying out of the PDF.
-        let cmap = create_cmap(font, glyph_set);
-        ctx.pdf.cmap(cmap_ref, &cmap.finish());
-
-        // Subset and write the font's bytes.
-        let glyphs: Vec<_> = glyph_set.keys().copied().collect();
-        let data = subset_font(font, &glyphs);
-
-        let mut stream = ctx.pdf.stream(data_ref, &data);
-        stream.filter(Filter::FlateDecode);
-        if is_cff {
-            stream.pair(Name(b"Subtype"), Name(b"CIDFontType0C"));
-        }
-
-        stream.finish();
-
-        let mut font_descriptor =
-            write_font_descriptor(&mut ctx.pdf, descriptor_ref, font, &base_font);
-        if is_cff {
-            font_descriptor.font_file3(data_ref);
-        } else {
-            font_descriptor.font_file2(data_ref);
-        }
-    }
-}
-
-/// Writes color fonts as Type3 fonts
-fn write_color_fonts(ctx: &mut PdfContext) {
-    let color_font_map = ctx.color_font_map.take_map();
-    for (font, color_font) in color_font_map {
-        // For each Type3 font that is part of this family…
-        for (font_index, subfont_id) in color_font.refs.iter().enumerate() {
-            // Allocate some IDs.
-            let cmap_ref = ctx.alloc.bump();
-            let descriptor_ref = ctx.alloc.bump();
-            let widths_ref = ctx.alloc.bump();
-            // And a map between glyph IDs and the instructions to draw this
-            // glyph.
-            let mut glyphs_to_instructions = Vec::new();
-
-            let start = font_index * 256;
-            let end = (start + 256).min(color_font.glyphs.len());
-            let glyph_count = end - start;
-            let subset = &color_font.glyphs[start..end];
-            let mut widths = Vec::new();
-
-            let scale_factor = font.ttf().units_per_em() as f32;
-
-            // Write the instructions for each glyph.
-            for color_glyph in subset {
-                let instructions_stream_ref = ctx.alloc.bump();
-                let width =
-                    font.advance(color_glyph.gid).unwrap_or(Em::new(0.0)).to_font_units();
-                widths.push(width);
-                // Create a fake page context for `write_frame`. We are only
-                // interested in the contents of the page.
-                let size = color_glyph.frame.size();
-                let mut page_ctx = PageContext::new(ctx, size);
-                page_ctx.bottom = size.y.to_f32();
-                page_ctx.content.start_color_glyph(width);
-                page_ctx.transform(
-                    // Make the Y axis go upwards, while preserving aspect ratio
-                    Transform::scale(Ratio::one(), -size.aspect_ratio())
-                        // Also move the origin to the top left corner
-                        .post_concat(Transform::translate(Abs::zero(), size.y)),
-                );
-                write_frame(&mut page_ctx, &color_glyph.frame);
-
-                // Retrieve the stream of the page and write it.
-                let stream = page_ctx.content.finish();
-                ctx.pdf.stream(instructions_stream_ref, &stream);
-
-                // Use this stream as instructions to draw the glyph.
-                glyphs_to_instructions.push(instructions_stream_ref);
+pub fn write_fonts(
+    context: &WithGlobalRefs,
+) -> SourceResult<(PdfChunk, HashMap<Font, Ref>)> {
+    let mut chunk = PdfChunk::new();
+    let mut out = HashMap::new();
+    context.resources.traverse(&mut |resources| {
+        for font in resources.fonts.items() {
+            if out.contains_key(font) {
+                continue;
             }
 
-            // Write the Type3 font object.
-            let mut pdf_font = ctx.pdf.type3_font(*subfont_id);
-            pdf_font.pair(Name(b"Resources"), ctx.type3_font_resources_ref);
-            pdf_font.bbox(color_font.bbox);
-            pdf_font.matrix([1.0 / scale_factor, 0.0, 0.0, 1.0 / scale_factor, 0.0, 0.0]);
-            pdf_font.first_char(0);
-            pdf_font.last_char((glyph_count - 1) as u8);
-            pdf_font.pair(Name(b"Widths"), widths_ref);
-            pdf_font.to_unicode(cmap_ref);
-            pdf_font.font_descriptor(descriptor_ref);
+            let type0_ref = chunk.alloc();
+            let cid_ref = chunk.alloc();
+            let descriptor_ref = chunk.alloc();
+            let cmap_ref = chunk.alloc();
+            let data_ref = chunk.alloc();
+            out.insert(font.clone(), type0_ref);
 
-            // Write the /CharProcs dictionary, that maps glyph names to
-            // drawing instructions.
-            let mut char_procs = pdf_font.char_procs();
-            for (gid, instructions_ref) in glyphs_to_instructions.iter().enumerate() {
-                char_procs
-                    .pair(Name(eco_format!("glyph{gid}").as_bytes()), *instructions_ref);
+            let glyph_set = resources.glyph_sets.get(font).unwrap();
+            let glyph_remapper = resources.glyph_remappers.get(font).unwrap();
+            let ttf = font.ttf();
+
+            // Do we have a TrueType or CFF font?
+            //
+            // FIXME: CFF2 must be handled differently and requires PDF 2.0
+            // (or we have to convert it to CFF).
+            let is_cff = ttf
+                .raw_face()
+                .table(CFF)
+                .or_else(|| ttf.raw_face().table(CFF2))
+                .is_some();
+
+            let base_font = base_font_name(font, glyph_set);
+            let base_font_type0 = if is_cff {
+                eco_format!("{base_font}-{IDENTITY_H}")
+            } else {
+                base_font.clone()
+            };
+
+            // Write the base font object referencing the CID font.
+            chunk
+                .type0_font(type0_ref)
+                .base_font(Name(base_font_type0.as_bytes()))
+                .encoding_predefined(Name(IDENTITY_H.as_bytes()))
+                .descendant_font(cid_ref)
+                .to_unicode(cmap_ref);
+
+            // Write the CID font referencing the font descriptor.
+            let mut cid = chunk.cid_font(cid_ref);
+            cid.subtype(if is_cff { CidFontType::Type0 } else { CidFontType::Type2 });
+            cid.base_font(Name(base_font.as_bytes()));
+            cid.system_info(SYSTEM_INFO);
+            cid.font_descriptor(descriptor_ref);
+            cid.default_width(0.0);
+            if !is_cff {
+                cid.cid_to_gid_map_predefined(Name(b"Identity"));
             }
-            char_procs.finish();
 
-            // Write the /Encoding dictionary.
-            let names = (0..glyph_count)
-                .map(|gid| eco_format!("glyph{gid}"))
+            // Extract the widths of all glyphs.
+            // `remapped_gids` returns an iterator over the old GIDs in their new sorted
+            // order, so we can append the widths as is.
+            let widths = glyph_remapper
+                .remapped_gids()
+                .map(|gid| {
+                    let width = ttf.glyph_hor_advance(GlyphId(gid)).unwrap_or(0);
+                    font.to_em(width).to_font_units()
+                })
                 .collect::<Vec<_>>();
-            pdf_font
-                .encoding_custom()
-                .differences()
-                .consecutive(0, names.iter().map(|name| Name(name.as_bytes())));
-            pdf_font.finish();
 
-            // Encode a CMAP to make it possible to search or copy glyphs.
-            let glyph_set = ctx.glyph_sets.get_mut(&font).unwrap();
-            let mut cmap = UnicodeCmap::new(CMAP_NAME, SYSTEM_INFO);
-            for (index, glyph) in subset.iter().enumerate() {
-                let Some(text) = glyph_set.get(&glyph.gid) else {
-                    continue;
-                };
-
-                if !text.is_empty() {
-                    cmap.pair_with_multiple(index as u8, text.chars());
+            // Write all non-zero glyph widths.
+            let mut first = 0;
+            let mut width_writer = cid.widths();
+            for (w, group) in widths.group_by_key(|&w| w) {
+                let end = first + group.len();
+                if w != 0.0 {
+                    let last = end - 1;
+                    width_writer.same(first as u16, last as u16, w);
                 }
+                first = end;
             }
-            ctx.pdf.cmap(cmap_ref, &cmap.finish());
 
-            // Write the font descriptor.
-            let postscript_name = font
-                .find_name(name_id::POST_SCRIPT_NAME)
-                .unwrap_or_else(|| "unknown".to_string());
-            let base_font = eco_format!("COLOR{font_index:x}+{postscript_name}");
-            write_font_descriptor(&mut ctx.pdf, descriptor_ref, &font, &base_font);
+            width_writer.finish();
+            cid.finish();
 
-            // Write the widths array
-            ctx.pdf.indirect(widths_ref).array().items(widths);
+            // Write the /ToUnicode character map, which maps glyph ids back to
+            // unicode codepoints to enable copying out of the PDF.
+            let cmap = create_cmap(glyph_set, glyph_remapper);
+            chunk
+                .cmap(cmap_ref, &cmap)
+                .writing_mode(WMode::Horizontal)
+                .filter(Filter::FlateDecode);
+
+            let subset = subset_font(font, glyph_remapper)
+                .map_err(|err| {
+                    let postscript_name = font.find_name(name_id::POST_SCRIPT_NAME);
+                    let name = postscript_name.as_deref().unwrap_or(&font.info().family);
+                    eco_format!("failed to process font {name}: {err}")
+                })
+                .at(Span::detached())?;
+
+            let mut stream = chunk.stream(data_ref, &subset);
+            stream.filter(Filter::FlateDecode);
+            if is_cff {
+                stream.pair(Name(b"Subtype"), Name(b"CIDFontType0C"));
+            }
+            stream.finish();
+
+            let mut font_descriptor =
+                write_font_descriptor(&mut chunk, descriptor_ref, font, &base_font);
+            if is_cff {
+                font_descriptor.font_file3(data_ref);
+            } else {
+                font_descriptor.font_file2(data_ref);
+            }
         }
-    }
+
+        Ok(())
+    })?;
+
+    Ok((chunk, out))
 }
 
 /// Writes a FontDescriptor dictionary.
-fn write_font_descriptor<'a>(
-    pdf: &'a mut pdf_writer::Pdf,
-    descriptor_ref: pdf_writer::Ref,
+pub fn write_font_descriptor<'a>(
+    pdf: &'a mut Chunk,
+    descriptor_ref: Ref,
     font: &'a Font,
-    base_font: &EcoString,
+    base_font: &str,
 ) -> FontDescriptor<'a> {
     let ttf = font.ttf();
     let metrics = font.metrics();
-    let postscript_name = font
+    let serif = font
         .find_name(name_id::POST_SCRIPT_NAME)
-        .unwrap_or_else(|| "unknown".to_string());
+        .is_some_and(|name| name.contains("Serif"));
 
     let mut flags = FontFlags::empty();
-    flags.set(FontFlags::SERIF, postscript_name.contains("Serif"));
+    flags.set(FontFlags::SERIF, serif);
     flags.set(FontFlags::FIXED_PITCH, ttf.is_monospaced());
     flags.set(FontFlags::ITALIC, ttf.is_italic());
     flags.insert(FontFlags::SYMBOLIC);
@@ -292,15 +203,19 @@ fn write_font_descriptor<'a>(
 
 /// Subset a font to the given glyphs.
 ///
-/// - For a font with TrueType outlines, this returns the whole OpenType font.
-/// - For a font with CFF outlines, this returns just the CFF font program.
+/// - For a font with TrueType outlines, this produces the whole OpenType font.
+/// - For a font with CFF outlines, this produces just the CFF font program.
+///
+/// In both cases, this returns the already compressed data.
 #[comemo::memoize]
 #[typst_macros::time(name = "subset font")]
-fn subset_font(font: &Font, glyphs: &[u16]) -> Arc<Vec<u8>> {
+fn subset_font(
+    font: &Font,
+    glyph_remapper: &GlyphRemapper,
+) -> Result<Arc<Vec<u8>>, subsetter::Error> {
     let data = font.data();
-    let profile = subsetter::Profile::pdf(glyphs);
-    let subsetted = subsetter::subset(data, font.index(), profile);
-    let mut data = subsetted.as_deref().unwrap_or(data);
+    let subset = subsetter::subset(data, font.index(), glyph_remapper)?;
+    let mut data = subset.as_ref();
 
     // Extract the standalone CFF font program if applicable.
     let raw = ttf_parser::RawFace::parse(data, 0).unwrap();
@@ -308,15 +223,34 @@ fn subset_font(font: &Font, glyphs: &[u16]) -> Arc<Vec<u8>> {
         data = cff;
     }
 
-    Arc::new(deflate(data))
+    Ok(Arc::new(deflate(data)))
+}
+
+/// Creates the base font name for a font with a specific glyph subset.
+/// Consists of a subset tag and the PostScript name of the font.
+///
+/// Returns a string of length maximum 116, so that even with `-Identity-H`
+/// added it does not exceed the maximum PDF/A name length of 127.
+pub(crate) fn base_font_name<T: Hash>(font: &Font, glyphs: &T) -> EcoString {
+    const MAX_LEN: usize = Name::PDFA_LIMIT - REST_LEN;
+    const REST_LEN: usize = SUBSET_TAG_LEN + 1 + 1 + IDENTITY_H.len();
+
+    let postscript_name = font.find_name(name_id::POST_SCRIPT_NAME);
+    let name = postscript_name.as_deref().unwrap_or("unknown");
+    let trimmed = &name[..name.len().min(MAX_LEN)];
+
+    // Hash the full name (we might have trimmed) and the glyphs to produce
+    // a fairly unique subset tag.
+    let subset_tag = subset_tag(&(name, glyphs));
+
+    eco_format!("{subset_tag}+{trimmed}")
 }
 
 /// Produce a unique 6 letter tag for a glyph set.
-fn subset_tag(glyphs: &BTreeMap<u16, EcoString>) -> EcoString {
-    const LEN: usize = 6;
+pub(crate) fn subset_tag<T: Hash>(glyphs: &T) -> EcoString {
     const BASE: u128 = 26;
-    let mut hash = typst::util::hash128(&glyphs);
-    let mut letter = [b'A'; LEN];
+    let mut hash = typst_utils::hash128(&glyphs);
+    let mut letter = [b'A'; SUBSET_TAG_LEN];
     for l in letter.iter_mut() {
         *l = b'A' + (hash % BASE) as u8;
         hash /= BASE;
@@ -324,72 +258,21 @@ fn subset_tag(glyphs: &BTreeMap<u16, EcoString>) -> EcoString {
     std::str::from_utf8(&letter).unwrap().into()
 }
 
-/// Create a /ToUnicode CMap.
-fn create_cmap(font: &Font, glyph_set: &mut BTreeMap<u16, EcoString>) -> UnicodeCmap {
-    let ttf = font.ttf();
-
-    // For glyphs that have codepoints mapping to them in the font's cmap table,
-    // we prefer them over pre-existing text mappings from the document. Only
-    // things that don't have a corresponding codepoint (or only a private-use
-    // one) like the "Th" in Linux Libertine get the text of their first
-    // occurrences in the document instead.
-    for subtable in ttf.tables().cmap.into_iter().flat_map(|table| table.subtables) {
-        if !subtable.is_unicode() {
-            continue;
-        }
-
-        subtable.codepoints(|n| {
-            let Some(c) = std::char::from_u32(n) else { return };
-            if c.general_category() == GeneralCategory::PrivateUse {
-                return;
-            }
-
-            let Some(GlyphId(g)) = ttf.glyph_index(c) else { return };
-            if glyph_set.contains_key(&g) {
-                glyph_set.insert(g, c.into());
-            }
-        });
-    }
-
+/// Create a compressed `/ToUnicode` CMap.
+#[comemo::memoize]
+#[typst_macros::time(name = "create cmap")]
+fn create_cmap(
+    glyph_set: &BTreeMap<u16, EcoString>,
+    glyph_remapper: &GlyphRemapper,
+) -> Arc<Vec<u8>> {
     // Produce a reverse mapping from glyphs' CIDs to unicode strings.
     let mut cmap = UnicodeCmap::new(CMAP_NAME, SYSTEM_INFO);
     for (&g, text) in glyph_set.iter() {
+        // See commend in `write_normal_text` for why we can choose the CID this way.
+        let cid = glyph_remapper.get(g).unwrap();
         if !text.is_empty() {
-            cmap.pair_with_multiple(glyph_cid(font, g), text.chars());
+            cmap.pair_with_multiple(cid, text.chars());
         }
     }
-
-    cmap
-}
-
-/// Get the CID for a glyph id.
-///
-/// When writing text into a PDF, we have to specify CIDs (character ids) not
-/// GIDs (glyph IDs).
-///
-/// Most of the time, the mapping between these two is an identity mapping. In
-/// particular, for TrueType fonts, the mapping is an identity mapping because
-/// of this line above:
-/// ```ignore
-/// cid.cid_to_gid_map_predefined(Name(b"Identity"));
-/// ```
-///
-/// However, CID-keyed CFF fonts may have a non-identity mapping defined in
-/// their charset. For those, we must map the glyph IDs in a `TextItem` to CIDs.
-/// The font defines the map through its charset. The charset usually maps
-/// glyphs to SIDs (string ids) specifying the glyph's name. Not for CID-keyed
-/// fonts though! For these, the SIDs are CIDs in disguise. Relevant quote from
-/// the CFF spec:
-///
-/// > The charset data, although in the same format as non-CIDFonts, will
-/// > represent CIDs rather than SIDs, [...]
-///
-/// This function performs the mapping from glyph ID to CID. It also works for
-/// non CID-keyed fonts. Then, it will simply return the glyph ID.
-pub(super) fn glyph_cid(font: &Font, glyph_id: u16) -> u16 {
-    font.ttf()
-        .tables()
-        .cff
-        .and_then(|cff| cff.glyph_cid(ttf_parser::GlyphId(glyph_id)))
-        .unwrap_or(glyph_id)
+    Arc::new(deflate(&cmap.finish()))
 }
