@@ -24,12 +24,6 @@ pub trait AstNode<'a>: Sized {
     }
 }
 
-/// A static syntax node used as a fallback value. This is returned instead of
-/// panicking when the syntactical structure isn't valid. In a normal
-/// compilation, evaluation isn't attempted on a broken file, but for IDE
-/// functionality, it is.
-static ARBITRARY: SyntaxNode = SyntaxNode::arbitrary();
-
 macro_rules! node {
     ($(#[$attr:meta])* $name:ident) => {
         #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
@@ -56,7 +50,9 @@ macro_rules! node {
         impl Default for $name<'_> {
             #[inline]
             fn default() -> Self {
-                Self(&ARBITRARY)
+                static PLACEHOLDER: SyntaxNode
+                    = SyntaxNode::placeholder(SyntaxKind::$name);
+                Self(&PLACEHOLDER)
             }
         }
     };
@@ -129,6 +125,8 @@ pub enum Expr<'a> {
     Math(Math<'a>),
     /// An identifier in math: `pi`.
     MathIdent(MathIdent<'a>),
+    /// A shorthand for a unicode codepoint in math: `a <= b`.
+    MathShorthand(MathShorthand<'a>),
     /// An alignment point in math: `&`.
     MathAlignPoint(MathAlignPoint<'a>),
     /// Matched delimiters in math: `[x + y]`.
@@ -236,6 +234,7 @@ impl<'a> AstNode<'a> for Expr<'a> {
             SyntaxKind::Equation => node.cast().map(Self::Equation),
             SyntaxKind::Math => node.cast().map(Self::Math),
             SyntaxKind::MathIdent => node.cast().map(Self::MathIdent),
+            SyntaxKind::MathShorthand => node.cast().map(Self::MathShorthand),
             SyntaxKind::MathAlignPoint => node.cast().map(Self::MathAlignPoint),
             SyntaxKind::MathDelimited => node.cast().map(Self::MathDelimited),
             SyntaxKind::MathAttach => node.cast().map(Self::MathAttach),
@@ -299,6 +298,7 @@ impl<'a> AstNode<'a> for Expr<'a> {
             Self::Equation(v) => v.to_untyped(),
             Self::Math(v) => v.to_untyped(),
             Self::MathIdent(v) => v.to_untyped(),
+            Self::MathShorthand(v) => v.to_untyped(),
             Self::MathAlignPoint(v) => v.to_untyped(),
             Self::MathDelimited(v) => v.to_untyped(),
             Self::MathAttach(v) => v.to_untyped(),
@@ -392,7 +392,7 @@ impl Expr<'_> {
 
 impl Default for Expr<'_> {
     fn default() -> Self {
-        Expr::Space(Space::default())
+        Expr::None(None::default())
     }
 }
 
@@ -454,7 +454,7 @@ node! {
 
 impl Shorthand<'_> {
     /// A list of all shorthands in markup mode.
-    pub const MARKUP_LIST: &'static [(&'static str, char)] = &[
+    pub const LIST: &'static [(&'static str, char)] = &[
         ("...", '…'),
         ("~", '\u{00A0}'),
         ("-", '\u{2212}'), // Only before a digit
@@ -463,52 +463,11 @@ impl Shorthand<'_> {
         ("-?", '\u{00AD}'),
     ];
 
-    /// A list of all shorthands in math mode.
-    pub const MATH_LIST: &'static [(&'static str, char)] = &[
-        ("...", '…'),
-        ("-", '\u{2212}'),
-        ("'", '′'),
-        ("*", '∗'),
-        ("!=", '≠'),
-        (":=", '≔'),
-        ("::=", '⩴'),
-        ("=:", '≕'),
-        ("<<", '≪'),
-        ("<<<", '⋘'),
-        (">>", '≫'),
-        (">>>", '⋙'),
-        ("<=", '≤'),
-        (">=", '≥'),
-        ("->", '→'),
-        ("-->", '⟶'),
-        ("|->", '↦'),
-        (">->", '↣'),
-        ("->>", '↠'),
-        ("<-", '←'),
-        ("<--", '⟵'),
-        ("<-<", '↢'),
-        ("<<-", '↞'),
-        ("<->", '↔'),
-        ("<-->", '⟷'),
-        ("~>", '⇝'),
-        ("~~>", '⟿'),
-        ("<~", '⇜'),
-        ("<~~", '⬳'),
-        ("=>", '⇒'),
-        ("|=>", '⤇'),
-        ("==>", '⟹'),
-        ("<==", '⟸'),
-        ("<=>", '⇔'),
-        ("<==>", '⟺'),
-        ("[|", '⟦'),
-        ("|]", '⟧'),
-        ("||", '‖'),
-    ];
-
     /// Get the shorthanded character.
     pub fn get(self) -> char {
         let text = self.0.text();
-        (Self::MARKUP_LIST.iter().chain(Self::MATH_LIST))
+        Self::LIST
+            .iter()
             .find(|&&(s, _)| s == text)
             .map_or_else(char::default, |&(_, c)| c)
     }
@@ -716,7 +675,7 @@ impl<'a> TermItem<'a> {
 }
 
 node! {
-    /// A mathemathical equation: `$x$`, `$ x^2 $`.
+    /// A mathematical equation: `$x$`, `$ x^2 $`.
     Equation
 }
 
@@ -771,6 +730,65 @@ impl Deref for MathIdent<'_> {
     /// may need to use [`get()`](Self::get) instead in some situations.
     fn deref(&self) -> &Self::Target {
         self.as_str()
+    }
+}
+
+node! {
+    /// A shorthand for a unicode codepoint in math: `a <= b`.
+    MathShorthand
+}
+
+impl MathShorthand<'_> {
+    /// A list of all shorthands in math mode.
+    pub const LIST: &'static [(&'static str, char)] = &[
+        ("...", '…'),
+        ("-", '−'),
+        ("'", '′'),
+        ("*", '∗'),
+        ("~", '∼'),
+        ("!=", '≠'),
+        (":=", '≔'),
+        ("::=", '⩴'),
+        ("=:", '≕'),
+        ("<<", '≪'),
+        ("<<<", '⋘'),
+        (">>", '≫'),
+        (">>>", '⋙'),
+        ("<=", '≤'),
+        (">=", '≥'),
+        ("->", '→'),
+        ("-->", '⟶'),
+        ("|->", '↦'),
+        (">->", '↣'),
+        ("->>", '↠'),
+        ("<-", '←'),
+        ("<--", '⟵'),
+        ("<-<", '↢'),
+        ("<<-", '↞'),
+        ("<->", '↔'),
+        ("<-->", '⟷'),
+        ("~>", '⇝'),
+        ("~~>", '⟿'),
+        ("<~", '⇜'),
+        ("<~~", '⬳'),
+        ("=>", '⇒'),
+        ("|=>", '⤇'),
+        ("==>", '⟹'),
+        ("<==", '⟸'),
+        ("<=>", '⇔'),
+        ("<==>", '⟺'),
+        ("[|", '⟦'),
+        ("|]", '⟧'),
+        ("||", '‖'),
+    ];
+
+    /// Get the shorthanded character.
+    pub fn get(self) -> char {
+        let text = self.0.text();
+        Self::LIST
+            .iter()
+            .find(|&&(s, _)| s == text)
+            .map_or_else(char::default, |&(_, c)| c)
     }
 }
 
@@ -1421,11 +1439,11 @@ pub enum BinOp {
     Assign,
     /// The containment operator: `in`.
     In,
-    /// The inversed containment operator: `not in`.
+    /// The inverse containment operator: `not in`.
     NotIn,
     /// The add-assign operator: `+=`.
     AddAssign,
-    /// The subtract-assign oeprator: `-=`.
+    /// The subtract-assign operator: `-=`.
     SubAssign,
     /// The multiply-assign operator: `*=`.
     MulAssign,
@@ -2044,9 +2062,26 @@ impl<'a> ImportItems<'a> {
     pub fn iter(self) -> impl DoubleEndedIterator<Item = ImportItem<'a>> {
         self.0.children().filter_map(|child| match child.kind() {
             SyntaxKind::RenamedImportItem => child.cast().map(ImportItem::Renamed),
-            SyntaxKind::Ident => child.cast().map(ImportItem::Simple),
+            SyntaxKind::ImportItemPath => child.cast().map(ImportItem::Simple),
             _ => Option::None,
         })
+    }
+}
+
+node! {
+    /// A path to a submodule's imported name: `a.b.c`.
+    ImportItemPath
+}
+
+impl<'a> ImportItemPath<'a> {
+    /// An iterator over the path's components.
+    pub fn iter(self) -> impl DoubleEndedIterator<Item = Ident<'a>> {
+        self.0.children().filter_map(SyntaxNode::cast)
+    }
+
+    /// The name of the imported item. This is the last segment in the path.
+    pub fn name(self) -> Ident<'a> {
+        self.iter().last().unwrap_or_default()
     }
 }
 
@@ -2055,18 +2090,26 @@ impl<'a> ImportItems<'a> {
 pub enum ImportItem<'a> {
     /// A non-renamed import (the item's name in the scope is the same as its
     /// name).
-    Simple(Ident<'a>),
+    Simple(ImportItemPath<'a>),
     /// A renamed import (the item was bound to a different name in the scope
     /// than the one it was defined as).
     Renamed(RenamedImportItem<'a>),
 }
 
 impl<'a> ImportItem<'a> {
+    /// The path to the imported item.
+    pub fn path(self) -> ImportItemPath<'a> {
+        match self {
+            Self::Simple(path) => path,
+            Self::Renamed(renamed_item) => renamed_item.path(),
+        }
+    }
+
     /// The original name of the imported item, at its source. This will be the
     /// equal to the bound name if the item wasn't renamed with 'as'.
     pub fn original_name(self) -> Ident<'a> {
         match self {
-            Self::Simple(name) => name,
+            Self::Simple(path) => path.name(),
             Self::Renamed(renamed_item) => renamed_item.original_name(),
         }
     }
@@ -2075,7 +2118,7 @@ impl<'a> ImportItem<'a> {
     /// name, if it was renamed; otherwise, it's just its original name.
     pub fn bound_name(self) -> Ident<'a> {
         match self {
-            Self::Simple(name) => name,
+            Self::Simple(path) => path.name(),
             Self::Renamed(renamed_item) => renamed_item.new_name(),
         }
     }
@@ -2087,9 +2130,14 @@ node! {
 }
 
 impl<'a> RenamedImportItem<'a> {
-    /// The original name of the imported item (`a` in `a as d`).
-    pub fn original_name(self) -> Ident<'a> {
+    /// The path to the imported item.
+    pub fn path(self) -> ImportItemPath<'a> {
         self.0.cast_first_match().unwrap_or_default()
+    }
+
+    /// The original name of the imported item (`a` in `a as d` or `c.b.a as d`).
+    pub fn original_name(self) -> Ident<'a> {
+        self.path().name()
     }
 
     /// The new name of the imported item (`d` in `a as d`).
@@ -2097,7 +2145,7 @@ impl<'a> RenamedImportItem<'a> {
         self.0
             .children()
             .filter_map(SyntaxNode::cast)
-            .nth(1)
+            .last()
             .unwrap_or_default()
     }
 }
@@ -2133,5 +2181,15 @@ impl<'a> FuncReturn<'a> {
     /// The expression to return.
     pub fn body(self) -> Option<Expr<'a>> {
         self.0.cast_last_match()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_expr_default() {
+        assert!(Expr::default().to_untyped().cast::<Expr>().is_some());
     }
 }
