@@ -4,6 +4,16 @@
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
 #[repr(u8)]
 pub enum SyntaxKind {
+    /// The end of token stream.
+    End,
+    /// An invalid sequence of characters.
+    Error,
+
+    /// A line comment: `// ...`.
+    LineComment,
+    /// A block comment: `/* ... */`.
+    BlockComment,
+
     /// The contents of a file or content block.
     Markup,
     /// Plain text without markup.
@@ -65,6 +75,8 @@ pub enum SyntaxKind {
     Math,
     /// An identifier in math: `pi`.
     MathIdent,
+    /// A shorthand for a unicode codepoint in math: `a <= b`.
+    MathShorthand,
     /// An alignment point in math: `&`.
     MathAlignPoint,
     /// Matched delimiters in math: `[x + y]`.
@@ -252,6 +264,8 @@ pub enum SyntaxKind {
     ModuleImport,
     /// Items to import from a module: `a, b, c`.
     ImportItems,
+    /// A path to an imported name from a submodule: `a.b.c`.
+    ImportItemPath,
     /// A renamed import item: `a as d`.
     RenamedImportItem,
     /// A module include: `include "chapter1.typ"`.
@@ -266,15 +280,6 @@ pub enum SyntaxKind {
     Destructuring,
     /// A destructuring assignment expression: `(x, y) = (1, 2)`.
     DestructAssignment,
-
-    /// A line comment: `// ...`.
-    LineComment,
-    /// A block comment: `/* ... */`.
-    BlockComment,
-    /// An invalid sequence of characters.
-    Error,
-    /// The end of token stream.
-    End,
 }
 
 impl SyntaxKind {
@@ -352,7 +357,7 @@ impl SyntaxKind {
     pub fn is_trivia(self) -> bool {
         matches!(
             self,
-            Self::Space | Self::Parbreak | Self::LineComment | Self::BlockComment
+            Self::LineComment | Self::BlockComment | Self::Space | Self::Parbreak
         )
     }
 
@@ -364,6 +369,10 @@ impl SyntaxKind {
     /// A human-readable name for the kind.
     pub fn name(self) -> &'static str {
         match self {
+            Self::End => "end of tokens",
+            Self::Error => "syntax error",
+            Self::LineComment => "line comment",
+            Self::BlockComment => "block comment",
             Self::Markup => "markup",
             Self::Text => "text",
             Self::Space => "space",
@@ -393,6 +402,7 @@ impl SyntaxKind {
             Self::Equation => "equation",
             Self::Math => "math",
             Self::MathIdent => "math identifier",
+            Self::MathShorthand => "math shorthand",
             Self::MathAlignPoint => "math alignment point",
             Self::MathDelimited => "delimited math",
             Self::MathAttach => "math attachments",
@@ -483,6 +493,7 @@ impl SyntaxKind {
             Self::ForLoop => "for-loop expression",
             Self::ModuleImport => "`import` expression",
             Self::ImportItems => "import items",
+            Self::ImportItemPath => "imported item path",
             Self::RenamedImportItem => "renamed import item",
             Self::ModuleInclude => "`include` expression",
             Self::LoopBreak => "`break` expression",
@@ -490,10 +501,6 @@ impl SyntaxKind {
             Self::FuncReturn => "`return` expression",
             Self::Destructuring => "destructuring pattern",
             Self::DestructAssignment => "destructuring assignment expression",
-            Self::LineComment => "line comment",
-            Self::BlockComment => "block comment",
-            Self::Error => "syntax error",
-            Self::End => "end of tokens",
         }
     }
 }
