@@ -5,11 +5,14 @@ use std::ops::{Index, IndexMut, Range};
 use ecow::{eco_format, EcoString};
 use unicode_math_class::MathClass;
 
-use crate::set::SyntaxSet;
-use crate::{ast, is_ident, is_newline, set, LexMode, Lexer, SyntaxKind, SyntaxNode};
+use crate::set::{syntax_set, SyntaxSet};
+use crate::{
+    ast, is_ident, is_newline, set, LexMode, Lexer, SyntaxError, SyntaxKind, SyntaxNode,
+};
 
 /// Parses a source file.
 pub fn parse(text: &str) -> SyntaxNode {
+    let _scope = typst_timing::TimingScope::new("parse");
     let mut p = Parser::new(text, 0, LexMode::Markup);
     markup(&mut p, true, 0, |_| false);
     p.finish().into_iter().next().unwrap()
@@ -17,6 +20,7 @@ pub fn parse(text: &str) -> SyntaxNode {
 
 /// Parses top-level code.
 pub fn parse_code(text: &str) -> SyntaxNode {
+    let _scope = typst_timing::TimingScope::new("parse code");
     let mut p = Parser::new(text, 0, LexMode::Code);
     let m = p.marker();
     p.skip();
@@ -27,6 +31,7 @@ pub fn parse_code(text: &str) -> SyntaxNode {
 
 /// Parses top-level math.
 pub fn parse_math(text: &str) -> SyntaxNode {
+    let _scope = typst_timing::TimingScope::new("parse math");
     let mut p = Parser::new(text, 0, LexMode::Math);
     math(&mut p, |_| false);
     p.finish().into_iter().next().unwrap()
@@ -58,11 +63,7 @@ fn markup(
             continue;
         }
 
-        if p.at_set(set::MARKUP_EXPR) {
-            markup_expr(p, &mut at_start);
-        } else {
-            p.unexpected();
-        }
+        markup_expr(p, &mut at_start);
     }
     p.wrap(m, SyntaxKind::Markup);
 }
@@ -90,11 +91,7 @@ pub(super) fn reparse_markup(
             continue;
         }
 
-        if p.at_set(set::MARKUP_EXPR) {
-            markup_expr(&mut p, at_start);
-        } else {
-            p.unexpected();
-        }
+        markup_expr(&mut p, at_start);
     }
     (p.balanced && p.current_start() == range.end).then(|| p.finish())
 }
@@ -138,7 +135,10 @@ fn markup_expr(p: &mut Parser, at_start: &mut bool) {
         | SyntaxKind::TermMarker
         | SyntaxKind::Colon => p.convert(SyntaxKind::Text),
 
-        _ => {}
+        _ => {
+            p.unexpected();
+            return; // Don't set `at_start`
+        }
     }
 
     *at_start = false;
@@ -146,28 +146,18 @@ fn markup_expr(p: &mut Parser, at_start: &mut bool) {
 
 /// Parses strong content: `*Strong*`.
 fn strong(p: &mut Parser) {
-    const END: SyntaxSet = SyntaxSet::new()
-        .add(SyntaxKind::Star)
-        .add(SyntaxKind::Parbreak)
-        .add(SyntaxKind::RightBracket);
-
     let m = p.marker();
     p.assert(SyntaxKind::Star);
-    markup(p, false, 0, |p| p.at_set(END));
+    markup(p, false, 0, |p| p.at_set(syntax_set!(Star, Parbreak, RightBracket)));
     p.expect_closing_delimiter(m, SyntaxKind::Star);
     p.wrap(m, SyntaxKind::Strong);
 }
 
 /// Parses emphasized content: `_Emphasized_`.
 fn emph(p: &mut Parser) {
-    const END: SyntaxSet = SyntaxSet::new()
-        .add(SyntaxKind::Underscore)
-        .add(SyntaxKind::Parbreak)
-        .add(SyntaxKind::RightBracket);
-
     let m = p.marker();
     p.assert(SyntaxKind::Underscore);
-    markup(p, false, 0, |p| p.at_set(END));
+    markup(p, false, 0, |p| p.at_set(syntax_set!(Underscore, Parbreak, RightBracket)));
     p.expect_closing_delimiter(m, SyntaxKind::Underscore);
     p.wrap(m, SyntaxKind::Emph);
 }
@@ -190,16 +180,11 @@ fn raw(p: &mut Parser) {
 
 /// Parses a section heading: `= Introduction`.
 fn heading(p: &mut Parser) {
-    const END: SyntaxSet = SyntaxSet::new()
-        .add(SyntaxKind::Label)
-        .add(SyntaxKind::RightBracket)
-        .add(SyntaxKind::Space);
-
     let m = p.marker();
     p.assert(SyntaxKind::HeadingMarker);
     whitespace_line(p);
     markup(p, false, usize::MAX, |p| {
-        p.at_set(END)
+        p.at_set(syntax_set!(Label, Space, RightBracket))
             && (!p.at(SyntaxKind::Space) || p.lexer.clone().next() == SyntaxKind::Label)
     });
     p.wrap(m, SyntaxKind::Heading);
@@ -227,14 +212,11 @@ fn enum_item(p: &mut Parser) {
 
 /// Parses an item in a term list: `/ Term: Details`.
 fn term_item(p: &mut Parser) {
-    const TERM_END: SyntaxSet =
-        SyntaxSet::new().add(SyntaxKind::Colon).add(SyntaxKind::RightBracket);
-
     let m = p.marker();
     p.assert(SyntaxKind::TermMarker);
     let min_indent = p.column(p.prev_end());
     whitespace_line(p);
-    markup(p, false, usize::MAX, |p| p.at_set(TERM_END));
+    markup(p, false, usize::MAX, |p| p.at_set(syntax_set!(Colon, RightBracket)));
     p.expect(SyntaxKind::Colon);
     whitespace_line(p);
     markup(p, false, min_indent, |p| p.at(SyntaxKind::RightBracket));
@@ -317,7 +299,7 @@ fn math_expr_prec(p: &mut Parser, min_prec: usize, stop: SyntaxKind) {
             }
         }
 
-        SyntaxKind::Text | SyntaxKind::Shorthand => {
+        SyntaxKind::Text | SyntaxKind::MathShorthand => {
             continuable = matches!(
                 math_class(p.current_text()),
                 None | Some(MathClass::Alphabetic)
@@ -391,11 +373,6 @@ fn math_expr_prec(p: &mut Parser, min_prec: usize, stop: SyntaxKind) {
             continue;
         }
 
-        // Separate primes and superscripts to different attachments.
-        if primed && p.current() == SyntaxKind::Hat {
-            p.wrap(m, SyntaxKind::MathAttach);
-        }
-
         let Some((kind, stop, assoc, mut prec)) = math_op(p.current()) else {
             // No attachments, so we need to wrap primes as attachment.
             if primed {
@@ -427,7 +404,7 @@ fn math_expr_prec(p: &mut Parser, min_prec: usize, stop: SyntaxKind) {
         math_expr_prec(p, prec, stop);
         math_unparen(p, m2);
 
-        if p.eat_if(SyntaxKind::Underscore) || (!primed && p.eat_if(SyntaxKind::Hat)) {
+        if p.eat_if(SyntaxKind::Underscore) || p.eat_if(SyntaxKind::Hat) {
             let m3 = p.marker();
             math_expr_prec(p, prec, SyntaxKind::End);
             math_unparen(p, m3);
@@ -633,6 +610,10 @@ fn code_exprs(p: &mut Parser, mut stop: impl FnMut(&Parser) -> bool) {
             code_expr(p);
             if !p.end() && !stop(p) && !p.eat_if(SyntaxKind::Semicolon) {
                 p.expected("semicolon or line break");
+                if p.at(SyntaxKind::Label) {
+                    p.hint("labels can only be applied in markup mode");
+                    p.hint("try wrapping your code in a markup block (`[ ]`)");
+                }
             }
         }
 
@@ -824,16 +805,11 @@ pub(super) fn reparse_block(text: &str, range: Range<usize>) -> Option<SyntaxNod
 
 /// Parses a code block: `{ let x = 1; x + 2 }`.
 fn code_block(p: &mut Parser) {
-    const END: SyntaxSet = SyntaxSet::new()
-        .add(SyntaxKind::RightBrace)
-        .add(SyntaxKind::RightBracket)
-        .add(SyntaxKind::RightParen);
-
     let m = p.marker();
     p.enter(LexMode::Code);
     p.enter_newline_mode(NewlineMode::Continue);
     p.assert(SyntaxKind::LeftBrace);
-    code(p, |p| p.at_set(END));
+    code(p, |p| p.at_set(syntax_set!(RightBrace, RightBracket, RightParen)));
     p.expect_closing_delimiter(m, SyntaxKind::RightBrace);
     p.exit();
     p.exit_newline_mode();
@@ -988,9 +964,22 @@ fn module_import(p: &mut Parser) {
         // imported at the same time.
         p.expect(SyntaxKind::Ident);
     }
-    if p.eat_if(SyntaxKind::Colon) && !p.eat_if(SyntaxKind::Star) {
-        import_items(p);
+
+    if p.eat_if(SyntaxKind::Colon) {
+        if p.at(SyntaxKind::LeftParen) {
+            let m1 = p.marker();
+            p.enter_newline_mode(NewlineMode::Continue);
+            p.assert(SyntaxKind::LeftParen);
+
+            import_items(p);
+
+            p.expect_closing_delimiter(m1, SyntaxKind::RightParen);
+            p.exit_newline_mode();
+        } else if !p.eat_if(SyntaxKind::Star) {
+            import_items(p);
+        }
     }
+
     p.wrap(m, SyntaxKind::ModuleImport);
 }
 
@@ -1003,6 +992,13 @@ fn import_items(p: &mut Parser) {
             p.unexpected();
         }
 
+        // Nested import path: `a.b.c`
+        while p.eat_if(SyntaxKind::Dot) {
+            p.expect(SyntaxKind::Ident);
+        }
+
+        p.wrap(item_marker, SyntaxKind::ImportItemPath);
+
         // Rename imported item.
         if p.eat_if(SyntaxKind::As) {
             p.expect(SyntaxKind::Ident);
@@ -1013,6 +1009,7 @@ fn import_items(p: &mut Parser) {
             p.expect(SyntaxKind::Comma);
         }
     }
+
     p.wrap(m, SyntaxKind::ImportItems);
 }
 
@@ -1094,13 +1091,17 @@ fn expr_with_paren(p: &mut Parser, atomic: bool) {
     if p.at(SyntaxKind::Arrow) {
         p.restore(checkpoint);
         params(p);
-        p.assert(SyntaxKind::Arrow);
+        if !p.expect(SyntaxKind::Arrow) {
+            return;
+        }
         code_expr(p);
         p.wrap(m, SyntaxKind::Closure);
     } else if p.at(SyntaxKind::Eq) && kind != SyntaxKind::Parenthesized {
         p.restore(checkpoint);
         destructuring_or_parenthesized(p, true, &mut HashSet::new());
-        p.assert(SyntaxKind::Eq);
+        if !p.expect(SyntaxKind::Eq) {
+            return;
+        }
         code_expr(p);
         p.wrap(m, SyntaxKind::DestructAssignment);
     } else {
@@ -1452,12 +1453,11 @@ fn pattern_leaf<'s>(
     seen: &mut HashSet<&'s str>,
     dupe: Option<&'s str>,
 ) {
-    if !p.at_set(set::PATTERN_LEAF) {
-        if p.current().is_keyword() {
-            p.eat_and_get().expected("pattern");
-        } else {
-            p.expected("pattern");
-        }
+    if p.current().is_keyword() {
+        p.eat_and_get().expected("pattern");
+        return;
+    } else if !p.at_set(set::PATTERN_LEAF) {
+        p.expected("pattern");
         return;
     }
 
@@ -1750,8 +1750,8 @@ impl<'s> Parser<'s> {
     fn save(&mut self) {
         let text = self.current_text();
         if self.at(SyntaxKind::Error) {
-            let message = self.lexer.take_error().unwrap();
-            self.nodes.push(SyntaxNode::error(message, text));
+            let error = self.lexer.take_error().unwrap();
+            self.nodes.push(SyntaxNode::error(error, text));
         } else {
             self.nodes.push(SyntaxNode::leaf(self.current, text));
         }
@@ -1764,7 +1764,7 @@ impl<'s> Parser<'s> {
     fn next_non_trivia(lexer: &mut Lexer<'s>) -> SyntaxKind {
         loop {
             let next = lexer.next();
-            // Loop is terminatable, because SyntaxKind::Eof is not a trivia.
+            // Loop is terminable, because SyntaxKind::End is not a trivia.
             if !next.is_trivia() {
                 break next;
             }
@@ -1828,8 +1828,17 @@ impl<'s> Parser<'s> {
     /// Produce an error that the given `thing` was expected at the position
     /// of the marker `m`.
     fn expected_at(&mut self, m: Marker, thing: &str) {
-        let error = SyntaxNode::error(eco_format!("expected {thing}"), "");
+        let error =
+            SyntaxNode::error(SyntaxError::new(eco_format!("expected {thing}")), "");
         self.nodes.insert(m.0, error);
+    }
+
+    /// Produce a hint.
+    fn hint(&mut self, hint: &str) {
+        let m = self.before_trivia();
+        if let Some(error) = self.nodes.get_mut(m.0 - 1) {
+            error.hint(hint);
+        }
     }
 
     /// Consume the next token (if any) and produce an error stating that it was
