@@ -6,7 +6,7 @@ use std::iter::zip;
 use std::ops::Range;
 use std::sync::Arc;
 
-use comemo::Prehashed;
+use typst_utils::LazyHash;
 
 use crate::reparser::reparse;
 use crate::{is_newline, parse, FileId, LinkedNode, Span, SyntaxNode, VirtualPath};
@@ -24,21 +24,22 @@ pub struct Source(Arc<Repr>);
 #[derive(Clone)]
 struct Repr {
     id: FileId,
-    text: Prehashed<String>,
-    root: Prehashed<SyntaxNode>,
+    text: LazyHash<String>,
+    root: LazyHash<SyntaxNode>,
     lines: Vec<Line>,
 }
 
 impl Source {
     /// Create a new source file.
     pub fn new(id: FileId, text: String) -> Self {
+        let _scope = typst_timing::TimingScope::new("create source");
         let mut root = parse(&text);
         root.numberize(id, Span::FULL).unwrap();
         Self(Arc::new(Repr {
             id,
             lines: lines(&text),
-            text: Prehashed::new(text),
-            root: Prehashed::new(root),
+            text: LazyHash::new(text),
+            root: LazyHash::new(root),
         }))
     }
 
@@ -75,6 +76,7 @@ impl Source {
     ///
     /// Returns the range in the new source that was ultimately reparsed.
     pub fn replace(&mut self, new: &str) -> Range<usize> {
+        let _scope = typst_timing::TimingScope::new("replace source");
         let old = self.text();
 
         let mut prefix =
@@ -117,7 +119,7 @@ impl Source {
         let inner = Arc::make_mut(&mut self.0);
 
         // Update the text itself.
-        inner.text.update(|text| text.replace_range(replace.clone(), with));
+        inner.text.replace_range(replace.clone(), with);
 
         // Remove invalidated line starts.
         inner.lines.truncate(line + 1);
@@ -135,9 +137,7 @@ impl Source {
         ));
 
         // Incrementally reparse the replaced range.
-        inner
-            .root
-            .update(|root| reparse(root, &inner.text, replace, with.len()))
+        reparse(&mut inner.root, &inner.text, replace, with.len())
     }
 
     /// Get the length of the file in UTF-8 encoded bytes.
