@@ -1,18 +1,17 @@
 use std::fmt::{self, Debug, Formatter};
 use std::ops::Range;
 
-use comemo::Prehashed;
 use ecow::EcoString;
 use heck::{ToKebabCase, ToTitleCase};
 use pulldown_cmark as md;
 use serde::{Deserialize, Serialize};
 use typed_arena::Arena;
-use typst::diag::{FileResult, StrResult};
-use typst::eval::Tracer;
+use typst::diag::{FileError, FileResult, StrResult};
 use typst::foundations::{Bytes, Datetime};
 use typst::layout::{Abs, Point, Size};
 use typst::syntax::{FileId, Source, VirtualPath};
 use typst::text::{Font, FontBook};
+use typst::utils::LazyHash;
 use typst::{Library, World};
 use unscanny::Scanner;
 use yaml_front_matter::YamlFrontMatter;
@@ -27,9 +26,11 @@ pub struct Html {
     #[serde(skip)]
     md: String,
     #[serde(skip)]
-    description: Option<EcoString>,
-    #[serde(skip)]
     outline: Vec<OutlineItem>,
+    #[serde(skip)]
+    title: Option<EcoString>,
+    #[serde(skip)]
+    description: Option<EcoString>,
 }
 
 impl Html {
@@ -38,8 +39,9 @@ impl Html {
         Self {
             md: String::new(),
             raw,
-            description: None,
             outline: vec![],
+            title: None,
+            description: None,
         }
     }
 
@@ -48,10 +50,12 @@ impl Html {
     pub fn markdown(resolver: &dyn Resolver, md: &str, nesting: Option<usize>) -> Self {
         let mut text = md;
         let mut description = None;
+        let mut title = None;
         let document = YamlFrontMatter::parse::<Metadata>(md);
         if let Ok(document) = &document {
             text = &document.content;
-            description = Some(document.metadata.description.clone())
+            title = document.metadata.title.clone();
+            description = document.metadata.description.clone();
         }
 
         let options = md::Options::ENABLE_TABLES
@@ -98,8 +102,9 @@ impl Html {
         Html {
             md: text.into(),
             raw,
-            description,
             outline: handler.outline,
+            title,
+            description,
         }
     }
 
@@ -113,21 +118,23 @@ impl Html {
         &self.md
     }
 
+    /// The outline of the HTML.
+    pub fn outline(&self) -> Vec<OutlineItem> {
+        self.outline.clone()
+    }
+
     /// The title of the HTML.
     ///
     /// Returns `None` if the HTML doesn't start with an `h1` tag.
     pub fn title(&self) -> Option<&str> {
-        let mut s = Scanner::new(&self.raw);
-        s.eat_if("<h1").then(|| {
-            s.eat_until('>');
-            s.eat_if('>');
-            s.eat_until("</h1>")
+        self.title.as_deref().or_else(|| {
+            let mut s = Scanner::new(&self.raw);
+            s.eat_if("<h1").then(|| {
+                s.eat_until('>');
+                s.eat_if('>');
+                s.eat_until("</h1>")
+            })
         })
-    }
-
-    /// The outline of the HTML.
-    pub fn outline(&self) -> Vec<OutlineItem> {
-        self.outline.clone()
     }
 
     /// The description from the front matter.
@@ -145,7 +152,8 @@ impl Debug for Html {
 /// Front matter metadata.
 #[derive(Deserialize)]
 struct Metadata {
-    description: EcoString,
+    title: Option<EcoString>,
+    description: Option<EcoString>,
 }
 
 struct Handler<'a> {
@@ -411,8 +419,7 @@ fn code_block(resolver: &dyn Resolver, lang: &str, text: &str) -> Html {
     let source = Source::new(id, compile);
     let world = DocWorld(source);
 
-    let mut tracer = Tracer::new();
-    let mut document = match typst::compile(&world, &mut tracer) {
+    let mut document = match typst::compile(&world).output {
         Ok(doc) => doc,
         Err(err) => {
             let msg = &err[0].message;
@@ -429,7 +436,7 @@ fn code_block(resolver: &dyn Resolver, lang: &str, text: &str) -> Html {
         document.pages.truncate(1);
     }
 
-    let hash = typst::util::hash128(text);
+    let hash = typst::utils::hash128(&(lang, text));
     resolver.example(hash, highlighted, &document)
 }
 
@@ -457,20 +464,24 @@ fn nest_heading(level: &mut md::HeadingLevel, nesting: usize) {
 struct DocWorld(Source);
 
 impl World for DocWorld {
-    fn library(&self) -> &Prehashed<Library> {
+    fn library(&self) -> &LazyHash<Library> {
         &LIBRARY
     }
 
-    fn book(&self) -> &Prehashed<FontBook> {
+    fn book(&self) -> &LazyHash<FontBook> {
         &FONTS.0
     }
 
-    fn main(&self) -> Source {
-        self.0.clone()
+    fn main(&self) -> FileId {
+        self.0.id()
     }
 
-    fn source(&self, _: FileId) -> FileResult<Source> {
-        Ok(self.0.clone())
+    fn source(&self, id: FileId) -> FileResult<Source> {
+        if id == self.0.id() {
+            Ok(self.0.clone())
+        } else {
+            Err(FileError::NotFound(id.vpath().as_rootless_path().into()))
+        }
     }
 
     fn file(&self, id: FileId) -> FileResult<Bytes> {
