@@ -1,18 +1,16 @@
 #![no_main]
 
-use comemo::Prehashed;
 use libfuzzer_sys::fuzz_target;
 use typst::diag::{FileError, FileResult};
-use typst::eval::Tracer;
 use typst::foundations::{Bytes, Datetime};
 use typst::syntax::{FileId, Source};
 use typst::text::{Font, FontBook};
-use typst::visualize::Color;
+use typst::utils::LazyHash;
 use typst::{Library, World};
 
 struct FuzzWorld {
-    library: Prehashed<Library>,
-    book: Prehashed<FontBook>,
+    library: LazyHash<Library>,
+    book: LazyHash<FontBook>,
     font: Font,
     source: Source,
 }
@@ -23,8 +21,8 @@ impl FuzzWorld {
         let font = Font::new(Bytes::from_static(data), 0).unwrap();
         let book = FontBook::from_fonts([&font]);
         Self {
-            library: Prehashed::new(Library::default()),
-            book: Prehashed::new(book),
+            library: LazyHash::new(Library::default()),
+            book: LazyHash::new(book),
             font,
             source: Source::detached(text),
         }
@@ -32,24 +30,28 @@ impl FuzzWorld {
 }
 
 impl World for FuzzWorld {
-    fn library(&self) -> &Prehashed<Library> {
+    fn library(&self) -> &LazyHash<Library> {
         &self.library
     }
 
-    fn book(&self) -> &Prehashed<FontBook> {
+    fn book(&self) -> &LazyHash<FontBook> {
         &self.book
     }
 
-    fn main(&self) -> Source {
-        self.source.clone()
+    fn main(&self) -> FileId {
+        self.source.id()
     }
 
-    fn source(&self, src: FileId) -> FileResult<Source> {
-        Err(FileError::NotFound(src.vpath().as_rootless_path().into()))
+    fn source(&self, id: FileId) -> FileResult<Source> {
+        if id == self.source.id() {
+            Ok(self.source.clone())
+        } else {
+            Err(FileError::NotFound(id.vpath().as_rootless_path().into()))
+        }
     }
 
-    fn file(&self, src: FileId) -> FileResult<Bytes> {
-        Err(FileError::NotFound(src.vpath().as_rootless_path().into()))
+    fn file(&self, id: FileId) -> FileResult<Bytes> {
+        Err(FileError::NotFound(id.vpath().as_rootless_path().into()))
     }
 
     fn font(&self, _: usize) -> Option<Font> {
@@ -63,10 +65,9 @@ impl World for FuzzWorld {
 
 fuzz_target!(|text: &str| {
     let world = FuzzWorld::new(text);
-    let mut tracer = Tracer::new();
-    if let Ok(document) = typst::compile(&world, &mut tracer) {
+    if let Ok(document) = typst::compile(&world).output {
         if let Some(page) = document.pages.first() {
-            std::hint::black_box(typst_render::render(&page.frame, 1.0, Color::WHITE));
+            std::hint::black_box(typst_render::render(page, 1.0));
         }
     }
     comemo::evict(10);
