@@ -2,15 +2,16 @@
 
 mod args;
 mod collect;
+mod custom;
 mod logger;
 mod run;
 mod world;
 
 use std::path::Path;
+use std::sync::LazyLock;
 use std::time::Duration;
 
 use clap::Parser;
-use once_cell::sync::Lazy;
 use parking_lot::Mutex;
 use rayon::iter::{ParallelBridge, ParallelIterator};
 
@@ -18,7 +19,7 @@ use crate::args::{CliArguments, Command};
 use crate::logger::Logger;
 
 /// The parsed command line arguments.
-static ARGS: Lazy<CliArguments> = Lazy::new(CliArguments::parse);
+static ARGS: LazyLock<CliArguments> = LazyLock::new(CliArguments::parse);
 
 /// The directory where the test suite is located.
 const SUITE_PATH: &str = "tests/suite";
@@ -29,6 +30,9 @@ const STORE_PATH: &str = "tests/store";
 /// The directory where the reference images are stored.
 const REF_PATH: &str = "tests/ref";
 
+/// The file where the skipped tests are stored.
+const SKIP_PATH: &str = "tests/skip.txt";
+
 /// The maximum size of reference images that aren't marked as `// LARGE`.
 const REF_LIMIT: usize = 20 * 1024;
 
@@ -37,7 +41,8 @@ fn main() {
 
     match &ARGS.command {
         None => test(),
-        Some(Command::Clean) => std::fs::remove_dir_all(STORE_PATH).unwrap(),
+        Some(Command::Clean) => clean(),
+        Some(Command::Undangle) => undangle(),
     }
 }
 
@@ -72,14 +77,20 @@ fn test() {
         }
     };
 
-    let filtered = tests.len();
-    if filtered == 0 {
+    let selected = tests.len();
+    if ARGS.list {
+        for test in tests.iter() {
+            println!("{test}");
+        }
+        eprintln!("{selected} selected, {skipped} skipped");
+        return;
+    } else if selected == 0 {
         eprintln!("no test selected");
         return;
     }
 
     // Run the tests.
-    let logger = Mutex::new(Logger::new(filtered, skipped));
+    let logger = Mutex::new(Logger::new(selected, skipped));
     std::thread::scope(|scope| {
         let logger = &logger;
         let (sender, receiver) = std::sync::mpsc::channel();
@@ -87,7 +98,10 @@ fn test() {
         // Regularly refresh the logger in case we make no progress.
         scope.spawn(move || {
             while receiver.recv_timeout(Duration::from_millis(500)).is_err() {
-                logger.lock().refresh();
+                if !logger.lock().refresh() {
+                    eprintln!("tests seem to be stuck");
+                    std::process::exit(1);
+                }
             }
         });
 
@@ -95,7 +109,7 @@ fn test() {
         //
         // We use `par_bridge` instead of `par_iter` because the former
         // results in a stack overflow during PDF export. Probably related
-        // to `typst::util::Deferred` yielding.
+        // to `typst::utils::Deferred` yielding.
         tests.iter().par_bridge().for_each(|test| {
             logger.lock().start(test);
             let result = std::panic::catch_unwind(|| run::run(test));
@@ -108,5 +122,23 @@ fn test() {
     let passed = logger.into_inner().finish();
     if !passed {
         std::process::exit(1);
+    }
+}
+
+fn clean() {
+    std::fs::remove_dir_all(STORE_PATH).unwrap();
+}
+
+fn undangle() {
+    match crate::collect::collect() {
+        Ok(_) => eprintln!("no danging reference images"),
+        Err(errors) => {
+            for error in errors {
+                if error.message == "dangling reference image" {
+                    std::fs::remove_file(&error.pos.path).unwrap();
+                    eprintln!("✅ deleted {}", error.pos.path.display());
+                }
+            }
+        }
     }
 }
